% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsheatmap.R
\name{tsheatmap}
\alias{tsheatmap}
\title{Heatmap Plot of Timeseries}
\usage{
tsheatmap(data, tzone = "UTC", datetime_colname = "Datetime",
  data_colname = NULL, xlab = "ISO week", ylab = "",
  LegendTitle = "")
}
\arguments{
\item{data}{a data.frame with at least one column of POSIXct values.}

\item{tzone}{give the timezone format to be used. default = "UTC"}

\item{datetime_colname}{column name of the column with POSIXct values of the data.frame \code{data}}

\item{data_colname}{column name of the column with values to be plotted of the data.frame \code{data}}

\item{xlab}{x axis label of the plot}

\item{ylab}{y axis label of the plot}

\item{LegendTitle}{legend title of the plot}
}
\value{
a heatmap of the timeseries data as plotly object
<!--######################################################################%##
-->  # 1 Input Checks
<!--  98311be45ae642d3a52660856199c561 -->
}
\description{
This function creates a plotly object with the heatmap of the timeseries data.frame.
}
