% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsheatmap.R
\name{tsheatmap}
\alias{tsheatmap}
\title{Heatmap Plot of Timeseries}
\usage{
tsheatmap(
  data,
  data_colname = NULL,
  tzone = "UTC",
  datetime_colname = "Datetime",
  z_limits = NULL,
  x_label = "ISO week",
  y_label = "Day of the week",
  LegendTitle = ""
)
}
\arguments{
\item{data}{a data.frame with at least one column of POSIXct values.}

\item{data_colname}{column name of the column with values to be plotted of the data.frame \code{data}}

\item{tzone}{give the timezone format to be used. default = "UTC"}

\item{datetime_colname}{column name of the column with POSIXct values of the data.frame \code{data}}

\item{z_limits}{a vector of two numeric \code{c(min,max)} e.g \code{c(0,100)} to scale the heatmap colors.}

\item{x_label}{x axis label of the plot}

\item{y_label}{y axis label of the plot}

\item{LegendTitle}{legend title of the plot}
}
\value{
a heatmap of the timeseries data as plotly object
}
\description{
This function creates a plotly object with the heatmap of the timeseries data.frame.
}
